server/AI_Core/requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.2
langgraph>=0.0.40
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0
nest-asyncio>=1.5.0
typing-extensions>=4.0.0
python-dateutil>=2.8.0
fastapi>=0.104.0
uvicorn>=0.24.0